requirements.txt
PyYAML>=6.0
fastapi>=0.110
uvicorn>=0.29
